WIP split shared arena modules and extract godmode document logic
